Add mongodb, docker, JWT

+++ frontend/app.js
@@ -8,18 +8,24 @@ app.set("views", path.join(__dirname, "views"));
 app.set("view engine", "ejs");
 
 // Docker network service URLs
-const USER_API = "http://localhost:3001";
-const PRODUCT_API = "http://localhost:3002";
-const ORDER_API = "http://localhost:3003";
+const USER_API = "http://user-service:3001";
+const PRODUCT_API = "http://product-service:3002";
+const ORDER_API = "http://order-service:3003";
+
+// const USER_API = "http://localhost:3001/";
+// const PRODUCT_API = "http://localhost:3002/";
+// const ORDER_API = "http://localhost:3003/";
 
 app.use(express.json());
 app.use(express.urlencoded({ extended: false }));
 app.use(express.static(path.join(__dirname, "public")));
 
+//MỞ TRANG ĐĂNG NHẬP
 app.get("/", (req, res) => {
   res.render("login", { title: "Đăng nhập" });
 });
 
+//ĐĂNG NHẬP
 app.post("/", async (req, res) => {
   try {
     const loginData = req.body;
@@ -46,10 +52,12 @@ app.post("/", async (req, res) => {
   }
 });
 
+//MỞ TRANG ĐĂNG KÝ
 app.get("/register", (req, res) => {
   res.render("register", { title: "Đăng ký" });
 });
 
+//ĐĂNG KÝ
 app.post("/register", async (req, res) => {
   try {
     const registerData = req.body;
@@ -73,17 +81,35 @@ app.post("/register", async (req, res) => {
   }
 });
 
+//MỞ TRANG ORDER
 app.get("/order", async (req, res) => {
   res.render("order", { title: "Đặt hàng" });
 });
 
+//LẤY DANH SÁCH SẢN PHẨM CHO TRANG ORDER
 app.get("/api/order", async (req, res) => {
   try {
-    const response = await fetch(PRODUCT_API);
+    const auth = req.headers["authorization"];
+    if (!auth) {
+      return res.status(401).json({
+        message: "Không có Authorization",
+      });
+    }
+    const token = auth.split(" ")[1];
+    const response = await fetch(PRODUCT_API, {
+      method: "GET",
+      headers: {
+        Authorization: `Bearer ${token}`,
+        "Content-Type": "application/json",
+      },
+    });
+
     const data = await response.json();
+
     if (!response.ok) {
       return res.status(response.status).json(data);
     }
+
     res.status(200).json(data);
   } catch (error) {
     res.status(500).json({
@@ -93,13 +119,54 @@ app.get("/api/order", async (req, res) => {
   }
 });
 
+//LẤY THÔNG TIN USER ĐỂ ĐẶT HÀNG
+app.get("/me", async (req, res) => {
+  try {
+    const auth = req.headers["authorization"];
+    if (!auth) {
+      return res.status(401).json({
+        message: "Không có Authorization",
+      });
+    }
+    const token = auth.split(" ")[1];
+    const response = await fetch(USER_API + "/me", {
+      method: "GET",
+      headers: {
+        Authorization: `Bearer ${token}`,
+        "Content-Type": "application/json",
+      },
+    });
+
+    const data = await response.json();
+
+    if (!response.ok) {
+      return res.status(response.status).json(data);
+    }
+
+    res.status(200).json(data);
+  } catch (error) {
+    res.status(500).json({
+      message: "Không kết nối được User Service",
+      error: error.message,
+    });
+  }
+});
+
 //THÊM ĐƠN HÀNG
 app.post("/order", async (req, res) => {
   try {
+    const auth = req.headers["authorization"];
+    if (!auth) {
+      return res.status(401).json({
+        message: "Không có Authorization",
+      });
+    }
+    const token = auth.split(" ")[1];
     const orderData = req.body;
     const response = await fetch(ORDER_API, {
       method: "POST",
       headers: {
+        Authorization: `Bearer ${token}`,
         "Content-Type": "application/json",
       },
       body: JSON.stringify(orderData),
@@ -124,8 +191,20 @@ app.get("/history", async (req, res) => {
 //LẤY LỊCH SỬ ĐƠN HÀNG CHO TRANG HISTORY
 app.get("/api/history", async (req, res) => {
   try {
-    const userId = req.query.userId;
-    const response = await fetch(ORDER_API + `/history?userId=${userId}`);
+    const auth = req.headers["authorization"];
+    if (!auth) {
+      return res.status(401).json({
+        message: "Không có Authorization",
+      });
+    }
+    const token = auth.split(" ")[1];
+    const response = await fetch(ORDER_API + "/history", {
+      method: "GET",
+      headers: {
+        Authorization: `Bearer ${token}`,
+        "Content-Type": "application/json",
+      },
+    });
     const data = await response.json();
     if (!response.ok) {
       return res.status(response.status).json(data);
@@ -142,8 +221,21 @@ app.get("/api/history", async (req, res) => {
 //LẤY CHI TIẾT ĐƠN HÀNG CHO TRANG HISTORY
 app.get("/api/history/:orderId", async (req, res) => {
   try {
+    const auth = req.headers["authorization"];
+    if (!auth) {
+      return res.status(401).json({
+        message: "Không có Authorization",
+      });
+    }
+    const token = auth.split(" ")[1];
     const orderId = req.params.orderId;
-    const response = await fetch(ORDER_API + `/history/${orderId}`);
+    const response = await fetch(ORDER_API + `/history/${orderId}`, {
+      method: "GET",
+      headers: {
+        Authorization: `Bearer ${token}`,
+        "Content-Type": "application/json",
+      },
+    });
     const data = await response.json();
     if (!response.ok) {
       return res.status(response.status).json(data);
@@ -164,9 +256,28 @@ app.get("/management", async (req, res) => {
 //LẤY TẤT CẢ ĐƠN HÀNG CHO TRANG MANAGEMENT
 app.get("/api/management", async (req, res) => {
   try {
+    const auth = req.headers["authorization"];
+    if (!auth) {
+      return res.status(401).json({
+        message: "Không có Authorization",
+      });
+    }
+    const token = auth.split(" ")[1];
     const [productRes, orderRes] = await Promise.all([
-      fetch(PRODUCT_API),
-      fetch(ORDER_API),
+      fetch(PRODUCT_API, {
+        method: "GET",
+        headers: {
+          Authorization: `Bearer ${token}`,
+          "Content-Type": "application/json",
+        },
+      }),
+      fetch(ORDER_API, {
+        method: "GET",
+        headers: {
+          Authorization: `Bearer ${token}`,
+          "Content-Type": "application/json",
+        },
+      }),
     ]);
 
     const products = await productRes.json();
@@ -195,10 +306,18 @@ app.get("/api/management", async (req, res) => {
 //THÊM MỚI SẢN PHẨM
 app.post("/management", async (req, res) => {
   try {
+    const auth = req.headers["authorization"];
+    if (!auth) {
+      return res.status(401).json({
+        message: "Không có Authorization",
+      });
+    }
+    const token = auth.split(" ")[1];
     const productData = req.body;
     const response = await fetch(PRODUCT_API, {
       method: "POST",
       headers: {
+        Authorization: `Bearer ${token}`,
         "Content-Type": "application/json",
       },
       body: JSON.stringify(productData),
@@ -219,11 +338,19 @@ app.post("/management", async (req, res) => {
 //SỬA SẢN PHẨM
 app.put("/management/product/:id", async (req, res) => {
   try {
+    const auth = req.headers["authorization"];
+    if (!auth) {
+      return res.status(401).json({
+        message: "Không có Authorization",
+      });
+    }
+    const token = auth.split(" ")[1];
     const productData = req.body;
     const productId = req.params.id;
     const response = await fetch(PRODUCT_API + `/${productId}`, {
       method: "PUT",
       headers: {
+        Authorization: `Bearer ${token}`,
         "Content-Type": "application/json",
       },
       body: JSON.stringify(productData),
@@ -232,6 +359,7 @@ app.put("/management/product/:id", async (req, res) => {
     if (!response.ok) {
       return res.status(response.status).json(data);
     }
+    console.log(102);
     res.status(201).json(data);
   } catch (error) {
     res.status(500).json({
@@ -242,13 +370,21 @@ app.put("/management/product/:id", async (req, res) => {
 });
 
 //SỬA TÌNH TRẠNG ĐƠN HÀNG
-app.put("/management/order/:id", async (req, res) => {
+app.patch("/management/order/:id", async (req, res) => {
   try {
+    const auth = req.headers["authorization"];
+    if (!auth) {
+      return res.status(401).json({
+        message: "Không có Authorization",
+      });
+    }
+    const token = auth.split(" ")[1];
     const orderData = req.body;
     const orderId = req.params.id;
     const response = await fetch(ORDER_API + `/${orderId}`, {
-      method: "PUT",
+      method: "PATCH",
       headers: {
+        Authorization: `Bearer ${token}`,
         "Content-Type": "application/json",
       },
       body: JSON.stringify(orderData),
@@ -266,11 +402,22 @@ app.put("/management/order/:id", async (req, res) => {
   }
 });
 
+//XÓA SẢN PHẨM
 app.delete("/management/product/:id", async (req, res) => {
   try {
+    const auth = req.headers["authorization"];
+    if (!auth) {
+      return res.status(401).json({
+        message: "Không có Authorization",
+      });
+    }
+    const token = auth.split(" ")[1];
     const productId = req.params.id;
     const response = await fetch(PRODUCT_API + `/${productId}`, {
       method: "DELETE",
+      headers: {
+        Authorization: `Bearer ${token}`,
+      },
     });
     const data = await response.json();
     if (!response.ok) {
@@ -285,11 +432,22 @@ app.delete("/management/product/:id", async (req, res) => {
   }
 });
 
+//XÓA ĐƠN HÀNG
 app.delete("/management/order/:id", async (req, res) => {
   try {
+    const auth = req.headers["authorization"];
+    if (!auth) {
+      return res.status(401).json({
+        message: "Không có Authorization",
+      });
+    }
+    const token = auth.split(" ")[1];
     const orderId = req.params.id;
     const response = await fetch(ORDER_API + `/${orderId}`, {
       method: "DELETE",
+      headers: {
+        Authorization: `Bearer ${token}`,
+      },
     });
     const data = await response.json();
     if (!response.ok) {
